+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.0.8 - In progress
 ### Changed
 - Demonstrate Docker Compose reverse proxy.
+- Simplify Vitessce demo to just a scatterplot.
 
 ## [v0.0.7](https://github.com/hubmapconsortium/portal-ui/tree/v0.0.7) - 2020-01-08
 ### Added